Replace instanceId in cloudwatch log config with actual value.
Previously, the gitlab cloudwatch log config contained the string
"instanceId". That should be replaced with the actual ID of the
instance.

In this commit, we re-introduce the lines which looked up the
instance ID and replaced it within the log config.

Additionally, we fix a tabbing issue in logging.tpl.